requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Django==4.0.3
django-ckeditor==6.2.0
django-js-asset==2.0.0
fonttools==4.35.0
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
mysqlclient==2.1.0
nltk==3.7
numpy==1.23.2
packaging==21.3
pandas==1.4.3
Pillow==9.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.2.1
regex==2022.7.25
scikit-learn==1.1.2
scikit-surprise==1.1.1
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.64.0
tzdata==2021.5
